# maintence-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Maintence API",
        "description": "API Documentation for Maintenance System",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get All Orders (Chunked)",
                "description": "Retrieve orders in chunks of 10. Use 'page' parameter to navigate.",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Chunk number (default 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrderResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create New Order",
                "description": "Creates a new maintenance order for a customer.",
                "operationId": "11147bac13e660416c0add2aaa304309",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "model_id",
                                    "member_id",
                                    "customer_name"
                                ],
                                "properties": {
                                    "model_id": {
                                        "description": "ID of the device model",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "member_id": {
                                        "description": "ID of the technician assigned",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "customer_name": {
                                        "type": "string",
                                        "example": "Adel Kharma"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Screen is cracked, touch not working",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get Order Details",
                "description": "Retrieve full details of a specific order, including its current status and bills.",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/OrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Edit Order Details",
                "description": "Updates manual fields like notes and on_hand_price. Does NOT affect bills/parts.",
                "operationId": "6a7029f5510ca08e28a2d58a5fe19439",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_name": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    },
                                    "on_hand_price": {
                                        "description": "Manual Price Override",
                                        "type": "number"
                                    },
                                    "member_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete Order",
                "description": "Deletes an order. WARNING: This will automatically delete all associated bills due to cascade constraints.",
                "operationId": "303a7b797887b66c7790c4c187b54790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 15
                    },
                    "customer_name": {
                        "type": "string",
                        "example": "Adel Kharma"
                    },
                    "device_model": {
                        "type": "string",
                        "example": "Samsung S24"
                    },
                    "assigned_member": {
                        "type": "string",
                        "example": "Ibrahim"
                    },
                    "add_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-06T10:00:00Z"
                    },
                    "status": {
                        "description": "pending, started, or finished",
                        "type": "string",
                        "example": "pending"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Screen crack",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}